faster version of np.outer
