+++ py/specter/psf/gausshermite.py
@@ -16,7 +16,7 @@ from scipy import special as sp
 
 from astropy.io import fits
 from specter.psf import PSF
-from specter.util import TraceSet
+from specter.util import TraceSet, outer
 
 class GaussHermitePSF(PSF):
     """
@@ -197,10 +197,11 @@ class GaussHermitePSF(PSF):
         
         #- Create core PSF image
         core1 = np.zeros((ny, nx))
+        spot1 = np.empty_like(core1)
         for i in range(degx1+1):
             for j in range(degy1+1):
                 c1 = self.coeff['GH-{}-{}'.format(i,j)].eval(ispec, wavelength)
-                spot1 = np.outer(yfunc1[j], xfunc1[i])
+                outer(yfunc1[j], xfunc1[i], out=spot1)
                 core1 += c1 * spot1
         
         #- Zero out elements in the core beyond 3 sigma
@@ -217,7 +218,7 @@ class GaussHermitePSF(PSF):
         # core2 = np.zeros((ny, nx))
         # for i in range(degx2+1):
         #     for j in range(degy2+1):
-        #         spot2 = np.outer(yfunc2[j], xfunc2[i])
+        #         spot2 = outer(yfunc2[j], xfunc2[i])
         #         c2 = self.coeff['GH2-{}-{}'.format(i,j)].eval(ispec, wavelength)
         #         core2 += c2 * spot2
